requirements.txt
anyascii==0.3.1
asgiref==3.6.0
beautifulsoup4==4.11.2
bootstrap4==0.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.9
django-ckeditor==6.3.2
django-embed-video==1.4.8
django-filter==22.1
django-js-asset==2.0.0
django-modelcluster==6.0
django-permissionedforms==0.1
djangorestframework==3.14.0
django-taggit==3.1.0
django-treebeard==4.6.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
l18n==2021.3
lxml==4.9.2
metadata-toolbox==0.1.0.dev0
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
parse==1.19.0
Pillow==9.4.0
pip==23.0.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
setuptools==40.8.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
telepath==0.3
typing_extensions==4.5.0
urllib3==1.26.15
wagtail==4.2.1
wagtail-embed-videos==0.4.2
webencodings==0.5.1
Willow==1.4.1
zipp==3.15.0
